refactor: relocate and split aggregation test files alongside implementation files
